requirements.txt
openai>=0.27.0
numpy
Pillow
opencv-python
pytest
